chore(test): moving test to tests
